refactor: update ValidationError to accept ErrorOptions as second param
