+++ lib/fromZodError.test.ts
@@ -1,61 +1,8 @@
 import * as zod from 'zod';
 import { ZodError } from 'zod';
 
-import {
-  fromZodError,
-  fromZodIssue,
-  isValidationError,
-  isValidationErrorLike,
-  ValidationError,
-} from './ValidationError';
-
-describe('fromZodIssue()', () => {
-  test('handles zod.string() schema errors', () => {
-    const schema = zod.string().email();
-
-    try {
-      schema.parse('foobar');
-    } catch (err) {
-      if (err instanceof ZodError) {
-        const validationError = fromZodIssue(err.issues[0]);
-        expect(validationError).toBeInstanceOf(ValidationError);
-        expect(validationError.message).toMatchInlineSnapshot(
-          `"Validation error: Invalid email"`
-        );
-        expect(validationError.details).toMatchInlineSnapshot(`
-                  [
-                    {
-                      "code": "invalid_string",
-                      "message": "Invalid email",
-                      "path": [],
-                      "validation": "email",
-                    },
-                  ]
-              `);
-      }
-    }
-  });
-
-  test('respects `includePath` prop when set to `false`', () => {
-    const schema = zod.object({
-      name: zod.string().min(3, '"Name" must be at least 3 characters'),
-    });
-
-    try {
-      schema.parse({ name: 'jo' });
-    } catch (err) {
-      if (err instanceof ZodError) {
-        const validationError = fromZodIssue(err.issues[0], {
-          includePath: false,
-        });
-        expect(validationError).toBeInstanceOf(ValidationError);
-        expect(validationError.message).toMatchInlineSnapshot(
-          `"Validation error: "Name" must be at least 3 characters"`
-        );
-      }
-    }
-  });
-});
+import { fromZodError } from './fromZodError';
+import { ValidationError } from './ValidationError';
 
 describe('fromZodError()', () => {
   test('handles zod.string() schema errors', () => {
@@ -70,16 +17,16 @@ describe('fromZodError()', () => {
         expect(validationError.message).toMatchInlineSnapshot(
           `"Validation error: Invalid email"`
         );
-        expect(validationError.details).toMatchInlineSnapshot(`
-                  [
-                    {
-                      "code": "invalid_string",
-                      "message": "Invalid email",
-                      "path": [],
-                      "validation": "email",
-                    },
-                  ]
-              `);
+        expect(validationError.issues).toMatchInlineSnapshot(`
+            [
+              {
+                "code": "invalid_string",
+                "message": "Invalid email",
+                "path": [],
+                "validation": "email",
+              },
+            ]
+        `);
       }
     }
   });
@@ -102,32 +49,32 @@ describe('fromZodError()', () => {
         expect(validationError.message).toMatchInlineSnapshot(
           `"Validation error: Number must be greater than 0 at "id"; String must contain at least 2 character(s) at "name""`
         );
-        expect(validationError.details).toMatchInlineSnapshot(`
-                  [
-                    {
-                      "code": "too_small",
-                      "exact": false,
-                      "inclusive": false,
-                      "message": "Number must be greater than 0",
-                      "minimum": 0,
-                      "path": [
-                        "id",
-                      ],
-                      "type": "number",
-                    },
-                    {
-                      "code": "too_small",
-                      "exact": false,
-                      "inclusive": true,
-                      "message": "String must contain at least 2 character(s)",
-                      "minimum": 2,
-                      "path": [
-                        "name",
-                      ],
-                      "type": "string",
-                    },
-                  ]
-              `);
+        expect(validationError.issues).toMatchInlineSnapshot(`
+          [
+            {
+              "code": "too_small",
+              "exact": false,
+              "inclusive": false,
+              "message": "Number must be greater than 0",
+              "minimum": 0,
+              "path": [
+                "id",
+              ],
+              "type": "number",
+            },
+            {
+              "code": "too_small",
+              "exact": false,
+              "inclusive": true,
+              "message": "String must contain at least 2 character(s)",
+              "minimum": 2,
+              "path": [
+                "name",
+              ],
+              "type": "string",
+            },
+          ]
+      `);
       }
     }
   });
@@ -144,37 +91,37 @@ describe('fromZodError()', () => {
         expect(validationError.message).toMatchInlineSnapshot(
           `"Validation error: Expected number, received string at index 1; Expected number, received boolean at index 2; Expected integer, received float at index 3"`
         );
-        expect(validationError.details).toMatchInlineSnapshot(`
-                  [
-                    {
-                      "code": "invalid_type",
-                      "expected": "number",
-                      "message": "Expected number, received string",
-                      "path": [
-                        1,
-                      ],
-                      "received": "string",
-                    },
-                    {
-                      "code": "invalid_type",
-                      "expected": "number",
-                      "message": "Expected number, received boolean",
-                      "path": [
-                        2,
-                      ],
-                      "received": "boolean",
-                    },
-                    {
-                      "code": "invalid_type",
-                      "expected": "integer",
-                      "message": "Expected integer, received float",
-                      "path": [
-                        3,
-                      ],
-                      "received": "float",
-                    },
-                  ]
-              `);
+        expect(validationError.issues).toMatchInlineSnapshot(`
+            [
+              {
+                "code": "invalid_type",
+                "expected": "number",
+                "message": "Expected number, received string",
+                "path": [
+                  1,
+                ],
+                "received": "string",
+              },
+              {
+                "code": "invalid_type",
+                "expected": "number",
+                "message": "Expected number, received boolean",
+                "path": [
+                  2,
+                ],
+                "received": "boolean",
+              },
+              {
+                "code": "invalid_type",
+                "expected": "integer",
+                "message": "Expected integer, received float",
+                "path": [
+                  3,
+                ],
+                "received": "float",
+              },
+            ]
+        `);
       }
     }
   });
@@ -203,43 +150,43 @@ describe('fromZodError()', () => {
         expect(validationError.message).toMatchInlineSnapshot(
           `"Validation error: Number must be greater than 0 at "id"; Expected number, received string at "arr[1]"; String must contain at least 2 character(s) at "nestedObj.name""`
         );
-        expect(validationError.details).toMatchInlineSnapshot(`
-                  [
-                    {
-                      "code": "too_small",
-                      "exact": false,
-                      "inclusive": false,
-                      "message": "Number must be greater than 0",
-                      "minimum": 0,
-                      "path": [
-                        "id",
-                      ],
-                      "type": "number",
-                    },
-                    {
-                      "code": "invalid_type",
-                      "expected": "number",
-                      "message": "Expected number, received string",
-                      "path": [
-                        "arr",
-                        1,
-                      ],
-                      "received": "string",
-                    },
-                    {
-                      "code": "too_small",
-                      "exact": false,
-                      "inclusive": true,
-                      "message": "String must contain at least 2 character(s)",
-                      "minimum": 2,
-                      "path": [
-                        "nestedObj",
-                        "name",
-                      ],
-                      "type": "string",
-                    },
-                  ]
-              `);
+        expect(validationError.issues).toMatchInlineSnapshot(`
+            [
+              {
+                "code": "too_small",
+                "exact": false,
+                "inclusive": false,
+                "message": "Number must be greater than 0",
+                "minimum": 0,
+                "path": [
+                  "id",
+                ],
+                "type": "number",
+              },
+              {
+                "code": "invalid_type",
+                "expected": "number",
+                "message": "Expected number, received string",
+                "path": [
+                  "arr",
+                  1,
+                ],
+                "received": "string",
+              },
+              {
+                "code": "too_small",
+                "exact": false,
+                "inclusive": true,
+                "message": "String must contain at least 2 character(s)",
+                "minimum": 2,
+                "path": [
+                  "nestedObj",
+                  "name",
+                ],
+                "type": "string",
+              },
+            ]
+        `);
       }
     }
   });
@@ -263,7 +210,7 @@ describe('fromZodError()', () => {
         expect(validationError.message).toMatchInlineSnapshot(
           `"Validation error: Invalid literal value, expected "success" at "custom-path.status""`
         );
-        expect(validationError.details).toMatchInlineSnapshot(`
+        expect(validationError.issues).toMatchInlineSnapshot(`
           [
             {
               "code": "invalid_literal",
@@ -304,7 +251,7 @@ describe('fromZodError()', () => {
         expect(validationError.message).toMatchInlineSnapshot(
           `"Validation error: Invalid literal value, expected "success" at "status"; Required at "data", or Invalid literal value, expected "error" at "status""`
         );
-        expect(validationError.details).toMatchInlineSnapshot(`
+        expect(validationError.issues).toMatchInlineSnapshot(`
           [
             {
               "code": "invalid_union",
@@ -362,7 +309,7 @@ describe('fromZodError()', () => {
         expect(validationError.message).toMatchInlineSnapshot(
           `"Validation error: Required at "terms""`
         );
-        expect(validationError.details).toMatchInlineSnapshot(`
+        expect(validationError.issues).toMatchInlineSnapshot(`
           [
             {
               "code": "invalid_union",
@@ -420,7 +367,7 @@ describe('fromZodError()', () => {
         expect(validationError.message).toMatchInlineSnapshot(
           `"Validation error: Invalid literal value, expected "value1" at "prop1"; Invalid literal value, expected "value2" at "prop2""`
         );
-        expect(validationError.details).toMatchInlineSnapshot(`
+        expect(validationError.issues).toMatchInlineSnapshot(`
           [
             {
               "code": "invalid_literal",
@@ -464,7 +411,7 @@ describe('fromZodError()', () => {
         expect(validationError.message).toMatchInlineSnapshot(
           `"Validation error: Expected string, received number at "."; Expected string, received boolean at "./*""`
         );
-        expect(validationError.details).toMatchInlineSnapshot(`
+        expect(validationError.issues).toMatchInlineSnapshot(`
           [
             {
               "code": "invalid_type",
@@ -506,91 +453,5 @@ describe('fromZodError()', () => {
         );
       }
     }
-  });
-});
-
-describe('isValidationError()', () => {
-  test('returns true when argument is instance of ValidationError', () => {
-    expect(isValidationError(new ValidationError('foobar'))).toEqual(true);
-  });
-
-  test('returns false when argument is plain Error', () => {
-    expect(isValidationError(new Error('foobar'))).toEqual(false);
-  });
-
-  test('returns false when argument is not an Error', () => {
-    expect(isValidationError('foobar')).toEqual(false);
-    expect(isValidationError(123)).toEqual(false);
-    expect(
-      isValidationError({
-        message: 'foobar',
-      })
-    ).toEqual(false);
-  });
-});
-
-describe('isValidationErrorLike()', () => {
-  test('returns true when argument is an actual instance of ValidationError', () => {
-    expect(isValidationErrorLike(new ValidationError('foobar'))).toEqual(true);
-  });
-
-  test('returns true when argument resembles a ValidationError', () => {
-    const err = new Error('foobar');
-    err.name = 'ZodValidationError'; // force ZodValidationError
-
-    expect(isValidationErrorLike(err)).toEqual(true);
-  });
-
-  test('returns false when argument is generic Error', () => {
-    expect(isValidationErrorLike(new Error('foobar'))).toEqual(false);
-  });
-
-  test('returns false when argument is not an Error instance', () => {
-    expect(isValidationErrorLike('foobar')).toEqual(false);
-    expect(isValidationErrorLike(123)).toEqual(false);
-    expect(
-      isValidationErrorLike({
-        message: 'foobar',
-      })
-    ).toEqual(false);
-  });
-});
-
-describe('ValidationError', () => {
-  describe('constructor', () => {
-    test('accepts message with details', () => {
-      const err = new ValidationError('Invalid email coyote@acme', [
-        {
-          code: 'invalid_string',
-          message: 'Invalid email',
-          path: [],
-          validation: 'email',
-        },
-      ]);
-      expect(err.details).toMatchInlineSnapshot(`
-        [
-          {
-            "code": "invalid_string",
-            "message": "Invalid email",
-            "path": [],
-            "validation": "email",
-          },
-        ]
-      `);
-    });
-
-    test('treats details as optional', () => {
-      const err = new ValidationError('Invalid email coyote@acme');
-      expect(err.details).toStrictEqual([]);
-    });
-  });
-
-  describe('toString()', () => {
-    test('converts error to string', () => {
-      const error = new ValidationError('Invalid email coyote@acme');
-      expect(error.toString()).toMatchInlineSnapshot(
-        `"Invalid email coyote@acme"`
-      );
-    });
   });
 });
