+++ lib/index.ts
@@ -1,13 +1,15 @@
+export { ValidationError } from './ValidationError';
+export { isValidationError } from './isValidationError';
+export { isValidationErrorLike } from './isValidationErrorLike';
+export { errorMap } from './errorMap';
 export {
-  ValidationError,
-  toValidationError,
-  isValidationError,
-  isValidationErrorLike,
-  fromZodError,
   fromZodIssue,
-  type ZodError,
-  type ZodIssue,
-  type FromZodErrorOptions,
   type FromZodIssueOptions,
-  errorMap,
-} from './ValidationError';
+  type ZodIssue,
+} from './fromZodIssue';
+export {
+  fromZodError,
+  type FromZodErrorOptions,
+  type ZodError,
+} from './fromZodError';
+export { toValidationError } from './toValidationError';
