+++ lib/isValidationError.ts
@@ -1,0 +1,5 @@
+import { ValidationError } from './ValidationError';
+
+export function isValidationError(err: unknown): err is ValidationError {
+  return err instanceof ValidationError;
+}
