+++ lib/isValidationErrorLike.ts
@@ -1,0 +1,5 @@
+import { ValidationError } from './ValidationError';
+
+export function isValidationErrorLike(err: unknown): err is ValidationError {
+  return err instanceof Error && err.name === 'ZodValidationError';
+}
